Trekk ut model
