+++ apps/innbyggers-flate/package.json
@@ -22,6 +22,7 @@
     "@navikt/ds-react": "^6.3.6",
     "@navikt/nav-dekoratoren-moduler": "^2.1.6",
     "dayjs": "^1.11.10",
+    "deltaker-flate-model": "workspace:*",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.22.1",
