+++ apps/innbyggers-flate/src/components/demo-banner/DemoStatusInstillinger.tsx
@@ -1,14 +1,11 @@
 import { Select } from '@navikt/ds-react'
 import { useState } from 'react'
 import { useDeltakerContext } from '../../DeltakerContext.tsx'
-import {
-  DeltakerResponse,
-  DeltakerStatusType,
-  deltakerSchema
-} from '../../api/data/deltaker.ts'
+import { DeltakerResponse, deltakerSchema } from '../../api/data/deltaker.ts'
 import { useDeferredFetch } from '../../hooks/useDeferredFetch'
 import { deltakerBffApiBasePath, useMock } from '../../utils/environment-utils'
 import { getDeltakerStatusDisplayText } from '../../utils/utils.ts'
+import { DeltakerStatusType } from 'deltaker-flate-model'
 
 export const endreMockDeltakelseStatus = (
   nyStatus: DeltakerStatusType
