+++ apps/innbyggers-flate/src/pages/AvbruttUtkastPage.tsx
@@ -1,7 +1,7 @@
 import { BodyLong, Heading, List } from '@navikt/ds-react'
 import { EMDASH, hentTiltakNavnHosArrangørTekst } from '../utils/utils'
 import { useDeltakerContext } from '../DeltakerContext'
-import { Tiltakstype } from '../api/data/deltaker'
+import { Tiltakstype } from 'deltaker-flate-model'
 import { UtkastHeader } from '../components/UtkastHeader.tsx'
 
 export const AvbruttUtkastPage = () => {
