+++ apps/nav-veileders-flate/package.json
@@ -22,6 +22,7 @@
     "@navikt/ds-react": "^6.3.6",
     "@navikt/ds-tailwind": "^6.3.6",
     "dayjs": "^1.11.10",
+    "deltaker-flate-model": "workspace:*",
     "mockjs": "^1.1.0",
     "msw": "^2.2.13",
     "react": "^18.2.0",
