+++ apps/nav-veileders-flate/src/components/demo-banner/DemoStatusInstillinger.tsx
@@ -1,14 +1,11 @@
 import { Select } from '@navikt/ds-react'
 import { useState } from 'react'
-import {
-  DeltakerStatusType,
-  PameldingResponse,
-  pameldingSchema
-} from '../../api/data/pamelding'
+import { PameldingResponse, pameldingSchema } from '../../api/data/pamelding'
 import { deltakerBffApiBasePath, useMock } from '../../utils/environment-utils'
 import { usePameldingContext } from '../tiltak/PameldingContext'
 import { useDeferredFetch } from '../../hooks/useDeferredFetch'
 import { getDeltakerStatusDisplayText } from '../../utils/displayText'
+import { DeltakerStatusType } from 'deltaker-flate-model'
 
 export const endreMockDeltakelseStatus = (
   nyStatus: DeltakerStatusType
