+++ apps/nav-veileders-flate/src/components/tiltak/endre-deltakelse-modaler/AvsluttDeltakelseModal.tsx
@@ -7,10 +7,7 @@ import {
   Textarea,
   useDatepicker
 } from '@navikt/ds-react'
-import {
-  DeltakerStatusAarsakType,
-  PameldingResponse
-} from '../../../api/data/pamelding.ts'
+import { PameldingResponse } from '../../../api/data/pamelding.ts'
 import { useState } from 'react'
 import {
   DeferredFetchState,
@@ -33,6 +30,7 @@ import {
 } from '../../../api/data/endre-deltakelse-request.ts'
 import { ModalFooter } from '../../ModalFooter.tsx'
 import { ErrorPage } from '../../../pages/ErrorPage.tsx'
+import { DeltakerStatusAarsakType } from 'deltaker-flate-model'
 
 interface AvsluttDeltakelseModalProps {
   pamelding: PameldingResponse
