+++ apps/nav-veileders-flate/src/components/tiltak/endre-deltakelse-modaler/EndreSluttarsakModal.tsx
@@ -1,8 +1,5 @@
 import { Detail, Modal, Radio, RadioGroup, Textarea } from '@navikt/ds-react'
-import {
-  DeltakerStatusAarsakType,
-  PameldingResponse
-} from '../../../api/data/pamelding.ts'
+import { PameldingResponse } from '../../../api/data/pamelding.ts'
 import { useState } from 'react'
 import {
   DeferredFetchState,
@@ -19,6 +16,7 @@ import {
 } from '../../../api/data/endre-deltakelse-request.ts'
 import { ModalFooter } from '../../ModalFooter.tsx'
 import { ErrorPage } from '../../../pages/ErrorPage.tsx'
+import { DeltakerStatusAarsakType } from 'deltaker-flate-model'
 
 interface EndreSluttarsakModalProps {
   pamelding: PameldingResponse
