+++ packages/model/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "deltaker-flate-model",
+  "version": "0.0.1",
+  "exports": {
+    ".": "./src/deltaker.ts"
+  },
+  "dependencies": {
+    "typescript": "^5.4.4"
+  },
+  "devDependencies": {
+    "@amt-deltaker-flate/eslint-config": "workspace:*"
+  }
+}
